backend/requirements.txt
asgiref==3.6.0
bleach==5.0.1
Django==4.1.5
django-cors-headers==3.13.0
django-environ==0.9.0
django-summernote==0.8.20.0
djangorestframework==3.14.0
essential-generators==1.0
Pillow==9.4.0
psycopg2==2.9.5
psycopg2-binary==2.9.5
pytz==2022.7.1
six==1.16.0
sqlparse==0.4.3
webencodings==0.5.1
